+++ javascript/linked-list/.gitignore
