--- javascript/linked-list/__tests__/linked-list.test.js
@@ -1,10 +1,0 @@
-'use strict';
-
-// Require our linked list implementation
-const LinkedLilst = require('../index');
-
-describe('Linked List', () => {
-  it('works', () => {
-    expect(true).toBeTruthy();
-  });
-});
